// frontend/src/api/models/balanceAdjustment.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Splice API
 * Financial account management API
 * OpenAPI spec version: 1.0
 */
import type { BalanceAdjustmentFlowDirection } from './balanceAdjustmentFlowDirection'
import type { Money } from './money'

export interface BalanceAdjustment {
  accountId: string
  accountName: string
  flowDirection: BalanceAdjustmentFlowDirection
  currency: string
  deltaAmount: number
  startBalance: Money
  endBalance: Money
}